give a parameter for the search radius to fetch a list of neighbouring sources
This will be useful to know whether we should expect GALEX to be blended.

Perhaps it can also be used for image deblending of aperture photometry,
sources overlapping with the sources within ~1-2'' and their radius are fine,
but point-sources further away that are accidentally added into the aperture
should be subtracted. An alternative is to force images to be monotonically
declining from the center outwards.